게시판 폼이랑 수정 삭제 권한

+++ animal_care_web/community/views.py
@@ -16,7 +16,7 @@ def board(request):
     }
     return render(request, 'community/board.html', context)
 
-
+@login_required(login_url='accounts:login')
 def posting(request, pk):
     post = Post.objects.get(pk=pk)
     context = {
@@ -24,16 +24,19 @@ def posting(request, pk):
     }
     return render(request, 'community/post.html', context)
 
-@login_required  # 작동 안함
+
+@login_required(login_url='accounts:login')
 def postwrite(request):
-    if not request.user.is_authenticated:
-        return render(request, 'community/login.html')
+    print('user')
+    print(request.user)
     # save post
     if request.method == 'POST':
         form = PostForm(request.POST)
         if form.is_valid():
-            print('save')
-            post = form.save(user_id=request.user)  # user_id 붙이려면 commit=False 해야되는지도..
+            post = form.save(commit=False)
+            print(request.user)
+            post.user = request.user
+            post.save()
             return redirect('community:posting', pk=post.pk)
     # write post
     form = PostForm()
@@ -43,13 +46,19 @@ def postwrite(request):
     return render(request, 'community/postwrite.html', context)
 
 
+@login_required(login_url='accounts:login')
 def postupdate(request, pk):
     post = Post.objects.get(pk=pk)
+    # 작성자가 아니면 수정 권한이 없음
+    if post.user.id != request.user.id:
+        return redirect('community:posting', pk)
+    # 수정본 저장
     if request.method == "POST":
         form = PostForm(request.POST, instance=post)
         if form.is_valid():
             form.save()
             return redirect('community:posting', post.pk)
+    # 수정 화면
     form = PostForm(instance=post)
     context = {
         'form': form,
@@ -58,7 +67,11 @@ def postupdate(request, pk):
     return render(request, 'community/postupdate.html', context)
 
 
+@login_required(login_url='accounts:login')
 def delete(request, pk):
     post = Post.objects.get(pk=pk)
+    # 작성자가 아니면 삭제 권한이 없음
+    if post.user.id != request.user.id:
+        return redirect('community:posting', pk)
     post.delete()
     return redirect('community:board')
